LayoutTests/platform/glib/imported/w3c/web-platform-tests/editing/other/paste-clipboard-change.tentative_id=contenteditable-expected.txt

Original text

Original text

FAIL clipboard change during paste event handler assert_equals: expected "New text" but got "Original text"
